ai/requirements.txt
absl-py==1.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
astor==0.8.1
async-timeout==4.0.2
attrdict==2.0.1
attrs==23.1.0
Babel==2.12.1
bce-python-sdk==0.8.83
beautifulsoup4==4.12.2
blinker==1.6.2
boto3==1.15.18
botocore==1.18.18
Brotli==1.0.9
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cssselect==1.2.0
cssutils==2.6.0
cycler==0.11.0
Cython==0.29.34
decorator==5.1.1
dnspython==2.3.0
email-validator==2.0.0.post2
entrypoints==0.4
et-xmlfile==1.1.0
fastapi==0.95.1
ffmpy==0.3.0
filelock==3.12.0
fire==0.5.0
Flask==2.3.1
flask-babel==3.1.0
flatbuffers==23.3.3
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
ftfy==6.1.1
future==0.18.3
gevent==22.10.2
geventhttpclient==2.0.2
gluonnlp==0.10.0
google-auth==2.18.1
google-auth-oauthlib==1.0.0
gradio==3.27.0
gradio-client==0.1.3
graphviz==0.8.4
greenlet==2.0.2
grpcio==1.54.0
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
huggingface-hub==0.0.12
idna==2.6
imageio==2.28.0
imgaug==0.4.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
lazy-loader==0.2
linkify-it-py==2.0.0
lmdb==1.4.1
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mxnet==1.7.0.post2
networkx==3.1
numpy==1.23.1
oauthlib==3.2.2
onnx==1.13.1
onnxruntime==1.8.0
openai==0.27.4
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.8.10
packaging==23.1
paddle-bfloat==0.1.7
paddleocr==2.6.1.3
paddlepaddle==2.4.2
pandas==2.0.1
pdf2docx==0.5.6
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
premailer==3.10.0
protobuf==3.20.0
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclipper==1.3.0.post4
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7
pyDeprecate==0.3.1
pydub==0.25.1
PyMuPDF==1.20.2
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-multipart==0.0.6
python-rapidjson==1.10
pytorch-lightning==1.5.0
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
rapidfuzz==3.0.0
rarfile==4.0
regex==2023.3.23
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.3.7
sacremoses==0.0.53
scikit-image==0.20.0
scipy==1.10.1
semantic-version==2.10.0
sentencepiece==0.1.96
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.26.1
sympy==1.11.1
tensorboard==2.13.0
tensorboard-data-server==0.7.0
termcolor==2.3.0
tifffile==2023.4.12
tokenizers==0.10.3
toolz==0.12.0
torch==1.13.1
torchmetrics==0.11.4
torchvision==0.15.1
tqdm==4.65.0
transformers==4.10.1
tritonclient==2.33.0
typing-extensions==4.5.0
tzdata==2023.3
uc-micro-py==1.0.1
ujson==5.7.0
urllib3==1.25.11
uvicorn==0.21.1
visualdl==2.5.1
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.2
Werkzeug==2.3.0
x2paddle==1.4.1
yarl==1.9.2
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
